feat(debates): remove the per-claim Debate toggle
Second half of GEO-2740. geo-chat#82 made an indexed position imply readiness,
so the switch that used to opt in is now asking the server for something it has
already done. Net 2,421 lines out, 92 in.

Gone: `ClaimDebateReadiness`, `useClaimDebateReadiness`, `debate-readiness-intent`,
`ClaimReadinessToggle`, `ClaimDebateButton`, and the `useReadinessOnFirstPosition`
hook in the rematch picker that stood the viewer ready after they took a side —
the server does that now, from the response itself.

Most of the deleted weight is the optimistic-intent machine: a click recorded an
intent, the switch moved immediately, and the queue request was held until
geo-chat's copy of the response caught up. That existed because the toggle could
be pressed in the minute before a position indexed. Nothing presses it now.

Two things kept rather than deleted with it:

  - `useRetireConfirmedResponseIndexing`, extracted from the toggle component. It
    clears the optimistic response snapshot once geo-chat agrees, and that
    snapshot is what `useClaimResponseIndexedNotifier` watches to fire the
    notification that *creates* readiness after #82. It matters more than it did,
    not less. Wired into the two claim surfaces where `claim-response-summary`
    reads the snapshot for display; surfaces that never had a `DebateClaim` do
    not need it, and the notifier dedupes by `runId` anyway.
  - `useClaimReadinessByClaimId` in the rematch picker. The card still reports
    readiness even though nothing sets it by hand.

Dropping `ClaimDebateButton` also drops a `useDebateClaims` call per entity row,
which is the per-row geo-chat query GEO-2724 is about. Asserted in
`entity-row-actions.test.tsx` rather than left as an absence, since putting it
back would be expensive as well as wrong.

Tests: 23 were pinning the toggle. Those whose whole subject was the switch are
deleted; those with a surviving point — response batching, the publish gate, the
per-space readiness lookup — keep it and lose only the switch assertion. Two in
the rematch suite covered the auto-opt-in and go with it, including the one
pinning which space the opt-in bound to; that concern is the server's now, since
geo-chat resolves the claim in the space the response was indexed in.

339 files / 3628 tests pass. `tsc --noEmit` reports the same 6 pre-existing
errors as the branch point, none in touched files. `eslint` clean on every file
in the diff.
