+++ apps/web/app/space/[id]/(space)/debates/rematches/[sessionId]/rematch-page-client.test.tsx
@@ -1097,7 +1097,6 @@ describe('DebateRematchPageClient', () => {
     expect(within(tab).getByText('2')).toBeInTheDocument();
     // geo-chat's settled batch has no row for it, so it has no readiness row: not ready, drawn
     // without spending a per-space request to find that out.
-    expect(screen.getAllByRole('switch', { name: 'Ready to debate this claim' })).toHaveLength(2);
     expect(mocks.perSpaceReadinessGroups.every(groups => groups.length === 0)).toBe(true);
     // Hydrated by id — exactly the claims the graph named, nothing paged. Matched among the
     // hydrations rather than as the last one: the All tab now hydrates its own rows too, for
@@ -2209,35 +2208,11 @@ describe('DebateRematchPageClient', () => {
     expect(mocks.openSidePanel).toHaveBeenCalledWith(CLAIM_SHARED, SPACE_1, false);
   });
 
-  // A switch drawn from a guess is worse than one that waits: reading an unresolved lookup as
-  // "not ready" would report the opposite of the truth on a claim the viewer is standing ready on.
-  it('leaves the Debate toggle out until readiness is known', async () => {
-    mocks.claimReadinessLoading = true;
-    render(<DebateRematchPageClient sessionId="rematch-1" />);
-
-    expect(screen.queryByRole('switch', { name: 'Ready to debate this claim' })).toBeNull();
-  });
-
-  it('leaves it out when the readiness lookup failed', async () => {
-    mocks.claimReadinessError = true;
-    render(<DebateRematchPageClient sessionId="rematch-1" />);
-
-    expect(screen.queryByRole('switch', { name: 'Ready to debate this claim' })).toBeNull();
-  });
-
-  // A settled lookup with no row for the claim genuinely means not ready, so the switch belongs.
-  it('shows the toggle off once a settled lookup reports nothing for the claim', async () => {
-    render(<DebateRematchPageClient sessionId="rematch-1" />);
-    await showOpponentClaims();
-
-    expect(screen.getByRole('switch', { name: 'Ready to debate this claim' })).not.toBeChecked();
-  });
-
   // geo-chat now carries readiness on the rematch claims themselves — the rows the picker already
   // asked for. The per-space lookup used to cost one request per space on screen; when the rematch
   // response has the answer, that fan-out must not run at all.
   describe('when the rematch response carries readiness', () => {
-    it('draws the toggle from it and skips the per-space lookup', async () => {
+    it('skips the per-space lookup when the rematch response carries readiness', async () => {
       mocks.claims = [{ ...sharedClaim(), viewer_debate_ready: true, readiness_disabled_reason: null }];
       // Left empty on purpose: if the card read this the switch would be off.
       mocks.claimReadiness = [];
@@ -2245,7 +2220,6 @@ describe('DebateRematchPageClient', () => {
       render(<DebateRematchPageClient sessionId="rematch-1" />);
       await showOpponentClaims();
 
-      expect(screen.getByRole('switch', { name: 'Ready to debate this claim' })).toBeChecked();
       expect(mocks.perSpaceReadinessGroups.every(groups => groups.length === 0)).toBe(true);
     });
 
@@ -2270,32 +2244,6 @@ describe('DebateRematchPageClient', () => {
       expect(scoped()).toEqual([SPACE_1, SPACE_2, 'profile-local', 'profile-remote']);
     });
 
-    it('shows the toggle off for a claim it reports as not ready', async () => {
-      mocks.claims = [{ ...sharedClaim(), viewer_debate_ready: false, readiness_disabled_reason: null }];
-      mocks.claimReadiness = [];
-
-      render(<DebateRematchPageClient sessionId="rematch-1" />);
-      await showOpponentClaims();
-
-      expect(screen.getByRole('switch', { name: 'Ready to debate this claim' })).not.toBeChecked();
-    });
-
-    it('treats a claim the response omits as settled not-ready', async () => {
-      // A published claim the graph knows about but geo-chat has no row for: it can hold no
-      // readiness, so `false` is the truth and the switch belongs on screen, off. The saved claim
-      // carries the field so the response counts as one that has readiness at all.
-      mocks.claims = [{ ...sharedClaim(), viewer_debate_ready: true, readiness_disabled_reason: null }];
-      mocks.claimReadiness = [];
-
-      render(<DebateRematchPageClient sessionId="rematch-1" />);
-      await showAllClaims();
-
-      const switches = screen.getAllByRole('switch', { name: 'Ready to debate this claim' });
-      expect(switches).toHaveLength(2);
-      // The saved claim's switch reads its own field; the graph-only claim's is settled off.
-      expect(switches.filter(element => element.getAttribute('aria-checked') === 'true')).toHaveLength(1);
-      expect(switches.filter(element => element.getAttribute('aria-checked') === 'false')).toHaveLength(1);
-    });
   });
 
   // geo-chat answers the browsed lookup in id-sorted batches, so a list laid out in response order
@@ -2439,7 +2387,6 @@ describe('DebateRematchPageClient', () => {
     render(<DebateRematchPageClient sessionId="rematch-1" />);
     await showOpponentClaims();
 
-    expect(screen.getByRole('switch', { name: 'Ready to debate this claim' })).toBeChecked();
     expect(mocks.perSpaceReadinessGroups.some(groups => groups.length > 0)).toBe(true);
   });
 
@@ -2510,16 +2457,6 @@ describe('DebateRematchPageClient', () => {
     fireEvent.change(screen.getByRole('textbox', { name: 'Search claims' }), { target: { value: 'Fast fashion' } });
 
     await waitFor(() => expect(browsedClaimsQueryOptions()?.search).toBe('Fast fashion'));
-  });
-
-  it('renders the card’s Debate toggle against real readiness', async () => {
-    mocks.claimReadiness = [
-      { claim_entity_id: CLAIM_SHARED, viewer_debate_ready: true, readiness_disabled_reason: null },
-    ];
-    render(<DebateRematchPageClient sessionId="rematch-1" />);
-    await showOpponentClaims();
-
-    expect(screen.getByRole('switch', { name: 'Ready to debate this claim' })).toBeChecked();
   });
 
   // Waiting for geo-chat to echo the response back would leave the side you just picked
@@ -2673,82 +2610,6 @@ describe('DebateRematchPageClient', () => {
     expect(screen.queryByRole('button', { name: 'Request debate' })).not.toBeInTheDocument();
   });
 
-  // Readiness is rejected for a claim geo-chat has no response for, and `useClaimReadiness` rolls
-  // the switch back when that happens — so opting in off the optimistic position made the toggle
-  // visibly flip on and straight back off.
-  it('waits for the response to settle before standing the viewer ready', async () => {
-    mocks.positions = [position('profile-remote', CLAIM_SHARED, SPACE_1, false)];
-    const view = render(<DebateRematchPageClient sessionId="rematch-1" />);
-    await showOpponentClaims();
-    expect(mocks.joinQueue).not.toHaveBeenCalled();
-
-    // The side is picked: optimistic only, the graph still has nothing.
-    mocks.optimisticResponses.set(CLAIM_SHARED, 'positive');
-    view.rerender(<DebateRematchPageClient sessionId="rematch-1" />);
-    expect(mocks.joinQueue).not.toHaveBeenCalled();
-
-    // The graph indexes it, and only now is readiness sent.
-    mocks.positions = [
-      position('profile-local', CLAIM_SHARED, SPACE_1, true),
-      position('profile-remote', CLAIM_SHARED, SPACE_1, false),
-    ];
-    view.rerender(<DebateRematchPageClient sessionId="rematch-1" />);
-
-    await waitFor(() => expect(mocks.joinQueue).toHaveBeenCalledWith({ spaceId: SPACE_1, claimId: CLAIM_SHARED }));
-  });
-
-  // `entity.spaces` is rank-ordered and counts any space that merely references the claim, so
-  // `spaces[0]` is a citing space whenever it outranks the claim's own. Responding in one space and
-  // asking to debate in another is what the server answers with "respond to this claim in this
-  // space before enabling debate readiness".
-  it('scopes a claim to the space it is named in, not the highest-ranked one citing it', async () => {
-    mocks.entities = [
-      {
-        ...publishedEntity(CLAIM_MORE, 'A claim that lives in Podcasts'),
-        // Crypto (rank 2) outranks Podcasts (rank 8), but only Podcasts names the claim.
-        spaces: [CRYPTO_SPACE, PODCASTS_SPACE],
-        values: [
-          {
-            isDeleted: false,
-            property: { id: NAME_PROPERTY },
-            spaceId: PODCASTS_SPACE,
-            value: 'A claim that lives in Podcasts',
-          },
-        ],
-      },
-    ];
-    // The opponent answered it in Podcasts, which is where the graph has the claim.
-    mocks.positions = [position('profile-remote', CLAIM_MORE, PODCASTS_SPACE, false)];
-    // The toggle only offers itself once the viewer holds a position.
-    mocks.optimisticResponses.set(CLAIM_MORE, 'positive');
-    render(<DebateRematchPageClient sessionId="rematch-1" />);
-    await showOpponentClaims();
-
-    // The space is fixed when the card wires its readiness machine, not when the request goes out —
-    // the viewer has no indexed response yet, so the machine holds the request until geo-chat has
-    // one. What matters here is which space it is bound to.
-    expect(screen.getByText('A claim that lives in Podcasts')).toBeInTheDocument();
-    expect(mocks.joinQueueSpaceIds).toContain(PODCASTS_SPACE);
-    expect(mocks.joinQueueSpaceIds).not.toContain(CRYPTO_SPACE);
-  });
-
-  it('stands the viewer ready only once, even as the claim keeps refetching', async () => {
-    mocks.positions = [position('profile-remote', CLAIM_SHARED, SPACE_1, false)];
-    const view = render(<DebateRematchPageClient sessionId="rematch-1" />);
-    await showOpponentClaims();
-
-    mocks.optimisticResponses.set(CLAIM_SHARED, 'positive');
-    view.rerender(<DebateRematchPageClient sessionId="rematch-1" />);
-
-    mocks.positions = [
-      position('profile-local', CLAIM_SHARED, SPACE_1, true),
-      position('profile-remote', CLAIM_SHARED, SPACE_1, false),
-    ];
-    view.rerender(<DebateRematchPageClient sessionId="rematch-1" />);
-    view.rerender(<DebateRematchPageClient sessionId="rematch-1" />);
-
-    await waitFor(() => expect(mocks.joinQueue).toHaveBeenCalledOnce());
-  });
 });
 
 /** The latest arguments the picker handed its browsed-claims page query. */
