+++ apps/web/core/claims/browse/claim-page-view.tsx
@@ -6,8 +6,8 @@ import { TOPICS_PROPERTY_ID } from '~/core/claims/ontology';
 import { claimResponseKind } from '~/core/claims/response-kind';
 import { TAG_PROPERTY_ID } from '~/core/constants';
 import type { DebateClaim } from '~/core/debates/api';
-import { ClaimDebateReadiness } from '~/core/debates/claim-debate-readiness';
 import { useDebateActivity, useDebateClaims } from '~/core/debates/hooks';
+import { useRetireConfirmedResponseIndexing } from '~/core/debates/retire-confirmed-response-indexing';
 import { useCreateDebateRequest, useDebateRequests, useMatchmakingMatches } from '~/core/debates/matchmaking/hooks';
 import { HubPillButton } from '~/core/debates/matchmaking/hub-pill-button';
 import { PositionRow, useClaimPositionControl } from '~/core/debates/matchmaking/matchmaking-claim-card';
@@ -146,7 +146,6 @@ export function ClaimPageView({ entityId, spaceId }: { entityId: string; spaceId
           responseKind={responseKind}
           summary={summary}
           row={row}
-          isRowLoading={rowQuery.isLoading}
         />
 
         <ClaimDebates claimId={entityId} spaceId={spaceId} responseKind={responseKind} />
@@ -183,7 +182,6 @@ function ClaimPositionSection({
   responseKind,
   summary,
   row,
-  isRowLoading,
 }: {
   entityId: string;
   spaceId: string;
@@ -192,7 +190,6 @@ function ClaimPositionSection({
   summary: ClaimResponseSummary;
   /** geo-chat's row, or null — which for a claim nobody has answered is a settled answer. */
   row: DebateClaim | null;
-  isRowLoading: boolean;
 }) {
   const claim = React.useMemo(
     () => ({
@@ -227,24 +224,16 @@ function ClaimPositionSection({
   // restoration from being mistaken for a login somebody asked for.
   const promptSignIn = usePrivySignIn();
   const control = useClaimPositionControl({ claim, positions, readiness, onRequireSignIn: promptSignIn });
+  // See claims-page-client: retiring the optimistic snapshot outlived the toggle that used to own
+  // it, because `claim-response-summary` on this page reads that snapshot for display.
+  useRetireConfirmedResponseIndexing({ debateClaim: row, entityId, spaceId });
 
   return (
     <section aria-label="Your position" className="rounded-lg border border-grey-02 bg-white p-4 @[560px]:p-5">
-      {/* Label left, readiness switch right — the same header shape the hub's claim card uses, so
-          the control sits where someone who has used the panel already expects it. `items-start`
-          so the label stays put when the switch stacks an explanation beneath it. */}
       <div className="mb-2.5 flex items-start justify-between gap-3">
         <Text as="div" variant="metadataMedium" color="grey-04">
           Your position
         </Text>
-        <ClaimDebateReadiness
-          debateClaim={row}
-          entityId={entityId}
-          spaceId={spaceId}
-          canEnable={!row?.active_debate}
-          isLoading={isRowLoading}
-          compact
-        />
       </div>
       <PositionRow
         positions={control.optimisticPositions}
