+++ apps/web/core/debates/matchmaking/matches-tab.test.tsx
@@ -216,7 +216,6 @@ describe('MatchesTab', () => {
 
     expect(screen.getByRole('button', { name: /^Disagree/ })).toHaveAttribute('aria-pressed', 'true');
     expect(screen.queryByText('Respond to this claim to debate it.')).not.toBeInTheDocument();
-    expect(screen.getByRole('switch', { name: 'Ready to debate this claim' })).toBeEnabled();
     expect(mocks.resetIndexing).not.toHaveBeenCalled();
   });
 
@@ -254,17 +253,6 @@ describe('MatchesTab', () => {
     expect(screen.queryByRole('button', { name: /^Agree/ })).not.toBeInTheDocument();
     expect(screen.queryByRole('link', { name: 'Leftover fixture claim' })).not.toBeInTheDocument();
     expect(screen.getByText('Claim unavailable')).toBeInTheDocument();
-    // Readiness is geo-chat state, so it still works for a claim the graph can't resolve.
-    expect(screen.getByRole('switch', { name: 'Ready to debate this claim' })).toBeInTheDocument();
-  });
-
-  it('stands down from a claim by turning readiness off, never by clearing a response', async () => {
-    render(<MatchesTab onTabChange={vi.fn()} />);
-
-    fireEvent.click(screen.getByRole('switch', { name: 'Ready to debate this claim' }));
-
-    await waitFor(() => expect(mocks.leaveMutateAsync).toHaveBeenCalledWith({ claimId: CLAIM_ENTITY_ID }));
-    expect(mocks.submitResponse).not.toHaveBeenCalled();
   });
 
   // A request you send while marked unavailable could not be answered, so the design drops the
