+++ apps/web/core/debates/matchmaking/matchmaking-claim-card.test.tsx
@@ -309,42 +309,6 @@ describe('position avatar stack', () => {
 });
 
 describe('MatchmakingClaimCard', () => {
-  it('puts the debate toggle in the card header beside the space name', () => {
-    renderCard(<MatchmakingClaimCard claim={claim} positions={positions} readiness={readiness()} />);
-
-    const toggle = screen.getByRole('switch', { name: toggleName });
-    const spaceName = screen.getByText('Crypto');
-
-    // The header row is the toggle's own row: the space chip and the toggle share it, which is
-    // what puts the toggle top-right rather than below the response buttons.
-    const headerRow = toggle.parentElement?.parentElement;
-    expect(headerRow).not.toBeNull();
-    expect(headerRow?.contains(spaceName)).toBe(true);
-
-    // ...and the header sits above the claim, so the toggle precedes it in the document.
-    const claimText = screen.getByText(CLAIM_TEXT);
-    expect(toggle.compareDocumentPosition(claimText) & Node.DOCUMENT_POSITION_FOLLOWING).toBeTruthy();
-  });
-
-  it('keeps the toggle above the response buttons', () => {
-    renderCard(<MatchmakingClaimCard claim={claim} positions={positions} readiness={readiness()} />);
-
-    const toggle = screen.getByRole('switch', { name: toggleName });
-    const agree = screen.getByRole('button', { name: /^Agree/ });
-
-    expect(toggle.compareDocumentPosition(agree) & Node.DOCUMENT_POSITION_FOLLOWING).toBeTruthy();
-  });
-
-  it('still heads the card with the toggle when the claim is not on the graph', () => {
-    const offGraph = { ...claim, claim_entity_id: 'not-a-graph-id' };
-    renderCard(<MatchmakingClaimCard claim={offGraph} positions={positions} readiness={readiness()} />);
-
-    const toggle = screen.getByRole('switch', { name: toggleName });
-    expect(toggle.parentElement?.parentElement?.contains(screen.getByText('Crypto'))).toBe(true);
-    // The unavailable notice loses the toggle it used to sit beside, but must still be shown.
-    expect(screen.getByText('Claim unavailable')).toBeInTheDocument();
-  });
-
   // geo-chat owns the avatar stacks, and its copy trails the response by a publish, an index and a
   // notification. Filling the pill with your colour but not your face read as the response not
   // having counted.
@@ -470,18 +434,4 @@ describe('MatchmakingClaimCard', () => {
     expect(mocks.submitResponse).toHaveBeenCalled();
   });
 
-  it('leaves the toggle disabled and uncaptioned without a response', () => {
-    // The card still shows what to do: the response pills render directly beneath this header,
-    // so the disabled switch no longer carries a caption of its own.
-    renderCard(
-      <MatchmakingClaimCard
-        claim={claim}
-        positions={positions}
-        readiness={readiness({ viewer_response: null, viewer_debate_ready: false })}
-      />
-    );
-
-    expect(screen.getByRole('switch', { name: toggleName })).toBeDisabled();
-    expect(screen.queryByText('Respond to this claim to debate it.')).not.toBeInTheDocument();
-  });
 });
