+++ apps/web/partials/entity-page/entity-page-metadata-header.tsx
@@ -18,7 +18,6 @@ import {
   FORMAT_PROPERTY,
   RENDERABLE_TYPE_PROPERTY,
 } from '~/core/constants';
-import { ClaimDebateButton } from '~/core/debates/claim-debate-button';
 import { useCreateProperty } from '~/core/hooks/use-create-property';
 import { useUserIsEditing } from '~/core/hooks/use-user-is-editing';
 import { ID } from '~/core/id';
@@ -291,7 +290,6 @@ export function EntityPageMetadataHeader({ id, spaceId, isVoteable = false }: En
       </div>
       <div className="flex items-center gap-2">
         {isVoteable && <EntityVoteButtons entityId={id} spaceId={spaceId} />}
-        <ClaimDebateButton entityId={id} spaceId={spaceId} />
       </div>
     </div>
   );
