+++ apps/web/partials/entity-page/entity-row-actions.tsx
@@ -4,8 +4,6 @@ import * as React from 'react';
 
 import cx from 'classnames';
 
-import { ClaimDebateButton } from '~/core/debates/claim-debate-button';
-
 import { EntityVoteButtons } from './entity-vote-buttons';
 
 type Props = {
@@ -15,13 +13,12 @@ type Props = {
   className?: string;
 };
 
-/** Entity-row actions in the claim design order: response, supporting actions, Debate. */
+/** Entity-row actions in the claim design order: response, then supporting actions. */
 export function EntityRowActions({ entityId, spaceId, children, className }: Props) {
   return (
     <div className={cx('flex items-center gap-4', className)}>
       <EntityVoteButtons entityId={entityId} spaceId={spaceId} claimResponderAvatarsPosition="trailing" />
       {children}
-      <ClaimDebateButton entityId={entityId} spaceId={spaceId} />
     </div>
   );
 }
